create store from REDUX lib. ->createStore
