+++ package.json
@@ -19,7 +19,7 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "react-test-renderer": "17.0.2",
-    "redux": "4.1.2",
+    "redux": "^4.2.0",
     "redux-form": "8.3.8",
     "redux-thunk": "2.4.1",
     "reselect": "4.1.5",
