+++ src/Components/Dialogs/Dialogs.tsx
@@ -3,7 +3,7 @@ import d from "./Dialogs.module.css"
 //import {NavLink} from "react-router-dom";
 import Message from "./Message/Message";
 import DialogItem from "./DialogItem/DialogItem";
-import { updateNewMessageBodyCreator, sendMessageCreator } from '../../redux/state'
+import { updateNewMessageBodyCreator, sendMessageCreator } from '../../redux/dialogs-reducer'
 
 
 const Dialogs = (props: any) => {
