+++ src/index.tsx
@@ -1,5 +1,5 @@
 import './index.css';
-import store from "./redux/state";
+import store from "./redux/redux-store";
 import ReactDOM from "react-dom";
 import App from "./App";
 import { BrowserRouter } from "react-router-dom";
