+++ src/redux/dialogs-reducer.js
@@ -18,4 +18,15 @@ const dialogsReducer=(state, action)=>{
              
     }
 }
+
+export const updateNewMessageBodyCreator = (body) => ({
+    type: UPDATE_NEW_MESSAGE_BODY,
+    body:body,
+
+})
+export const sendMessageCreator = () => ({
+    type: SEND_MESSAGE,
+   
+})
+
 export default dialogsReducer
