+++ src/redux/redux-store.js
@@ -1,0 +1,5 @@
+import {createStore} from "redux"
+
+let store = createStore();
+
+export default store
