+++ src/redux/store.js
@@ -3,11 +3,6 @@ import profileReducer from "./profile-reducer";
 import dialogsReducer from "./dialogs-reducer";
 import sidebarReducer from "./sidebar-reducer";
 
-const ADD_POST ="ADD-POST";
-const UPDATE_NEW_POST_TEXT="UPDATE-NEW-POST-TEXT";
-const UPDATE_NEW_MESSAGE_BODY ="UPDATE-NEW-MESSAGE-BODY";
-const SEND_MESSAGE ="SEND-MESSAGE";
-
 let store = {
      _state : {
         profilePage: {
@@ -67,74 +62,5 @@ let store = {
     }
 }
 
-
-export const addPostsActionCreator = () => ({
-        type: ADD_POST,
-    
-})
-export const onPostChangeActionCreator = (text) => ({
-        type: UPDATE_NEW_POST_TEXT,
-        newText: text
-    
-})
-
-
-export const updateNewMessageBodyCreator = (body) => ({
-    type: UPDATE_NEW_MESSAGE_BODY,
-    body:body,
-
-})
-export const sendMessageCreator = () => ({
-    type: SEND_MESSAGE,
-   
-})
-
 export default store
 
-// let rerenderEntireTree=()=>{}
-
-// let state = {
-//     profilePage: {
-//         postData: [
-//             {message: "Hello, this is my 1 post through the props", id: v1(), likesCount: 23},
-//             {message: "Hello, this is my 2message post through the props", id: v1(), likesCount: 25},
-//             {message: "Hello, this is my 3 post through the props", id: v1(), likesCount: 27},
-//         ],newPostText:"FLUX data"
-//     },
-//     messagesPage: {
-//         dialogData: [
-//             {name: "John", id: v1()},
-//             {name: "Jerry", id: v1()},
-//             {name: "Andrea", id: v1()},
-//             {name: "Lois", id: v1()},
-//             {name: "Anatoliy", id: v1()},],
-//         messagesData: [
-//             {message: "Hello buddy", id: v1()},
-//             {message: "Go to the stroll today?", id: v1()},
-//             {message: "Do you learned something from React today ?", id: v1()},
-//         ]
-//     }
-// }
-
-// export let addPost = (postMessage) => {
-//     let newPost = {
-//         id: v1(),
-//         message: postMessage,
-//         likesCount: 0
-//     }
-//     store.state.profilePage.postData.push(newPost);
-//     store.state.profilePage.newPostText="";
-//     rerenderEntireTree(store.state);
-    
-    
-// }
-// export let updateNewPostText = (newText) => {
-   
-//     store.state.profilePage.newPostText=newText;
-//     rerenderEntireTree(store.state);
-    
-// }
-
-// export const subscribe=(observer)=>{
-//     rerenderEntireTree=observer;
-// }
